requirements.txt
google-generativeai
python-dotenv
duckduckgo-search
streamlit
watchdog